writeups/misc/Vovochka's_secret/src/app/requirements.txt
aiohttp==3.8.6
aiogram==2.23.1
psycopg2-binary==2.9.9